add validation on delete subvocational

+++ thesis/app/Http/Controllers/SubvocationalController.php
@@ -6,17 +6,18 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Http\Models\Subvocational;
 use App\Http\Models\Vocational;
+use App\Http\Models\SelectionSchedule;
 use Auth;
 
 class SubvocationalController extends Controller
 {
     /**
-    * Display a listing of the resource.
-    *
-    * @return \Illuminate\Http\Response
-    */
-   public function index(Request $request)
-   {
+     * Display a listing of the resource.
+     * 
+     * @return \Illuminate\Http\Response
+     */
+    public function index(Request $request)
+    {
         $role_id = Auth::user()->roleId();
         
         if ($role_id == 1) {
@@ -27,15 +28,15 @@ class SubvocationalController extends Controller
         } else {
             return redirect()->route('profile_users.show');
         }
-   }
+    }
 
-   /**
-    * Show the form for creating a new resource.
-    *
-    * @return \Illuminate\Http\Response
-    */
-   public function create()
-   {
+    /**
+     * Show the form for creating a new resource.
+     * 
+     * @return \Illuminate\Http\Response
+     */
+    public function create()
+    {
         $role_id = Auth::user()->roleId();
             
         if ($role_id == 1) {
@@ -45,16 +46,16 @@ class SubvocationalController extends Controller
         } else {
             return redirect()->route('profile_users.show');
         }
-   }
+    }
 
-   /**
-    * Store a newly created resource in storage.
-    *
-    * @param  \Illuminate\Http\Request  $request
-    * @return \Illuminate\Http\Response
-    */
-   public function store(Request $request)
-   {
+    /**
+     * Store a newly created resource in storage.
+     * 
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(Request $request)
+    {
         $this->validate($request, [
             'name' => 'required',
             'vocational_id' => 'required',
@@ -68,16 +69,16 @@ class SubvocationalController extends Controller
 
         return redirect()->route('subvocationals.index')
                         ->with('success','Sub-Kejuruan berhasil dibuat');
-   }
+    }
 
-   /**
-    * Display the specified resource.
-    *
-    * @param  int  $id
-    * @return \Illuminate\Http\Response
-    */
-   public function show($id)
-   {
+    /**
+     * Display the specified resource.
+     * 
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function show($id)
+    {
         $role_id = Auth::user()->roleId();
             
         if ($role_id == 1) {
@@ -87,16 +88,16 @@ class SubvocationalController extends Controller
         } else {
             return redirect()->route('profile_users.show');
         }
-   }
+    }
 
-   /**
-    * Show the form for editing the specified resource.
-    *
-    * @param  int  $id
-    * @return \Illuminate\Http\Response
-    */
-   public function edit($id)
-   {
+    /**
+     * Show the form for editing the specified resource.
+     * 
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function edit($id)
+    {
         $role_id = Auth::user()->roleId();
             
         if ($role_id == 1) {
@@ -108,17 +109,17 @@ class SubvocationalController extends Controller
         } else {
             return redirect()->route('profile_users.show');
         }
-   }
+    }
 
-   /**
-    * Update the specified resource in storage.
-    *
-    * @param  \Illuminate\Http\Request  $request
-    * @param  int  $id
-    * @return \Illuminate\Http\Response
-    */
-   public function update(Request $request, $id)
-   {
+    /**
+     * Update the specified resource in storage.
+     * 
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, $id)
+    {
         $this->validate($request, [
             'name' => 'required',
             'vocational_id' => 'required',
@@ -133,19 +134,28 @@ class SubvocationalController extends Controller
 
         return redirect()->route('subvocationals.index')
                         ->with('success','Sub-Kejuruan berhasil diedit');
-   }
+    }
 
-   /**
-    * Remove the specified resource from storage.
-    *
-    * @param  int  $id
-    * @return \Illuminate\Http\Response
-    */
-   public function destroy($id)
-   {
-        Subvocational::find($id)->delete();
+    /**
+     * Remove the specified resource from storage.
+     * 
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy($id)
+    {
+        $selectionschedule = SelectionSchedule::with('subvocational')
+                                                ->where('sub_vocational_id', '=', $id)
+                                                ->first();
         
-        return redirect()->route('subvocationals.index')
-                        ->with('success','Sub-Kejuruan berhasil dihapus');
-   }
+        if ($selectionschedule == null) {
+            Subvocational::find($id)->delete();
+
+            return redirect()->route('subvocationals.index')
+                             ->with('success','Sub-Kejuruan berhasil dihapus');
+        } else {
+            return redirect()->route('subvocationals.index')
+                             ->with('failed','Sub-Kejuruan tidak bisa dihapus karena sudah memiliki jadwal seleksi');
+        } 
+    }
 }
